Create Plus Minus
Given an array of integers, calculate the ratios of its elements that are positive, negative, and zero. Print the decimal value of each fraction on a new line with  places after the decimal.

Note: This challenge introduces precision problems. The test cases are scaled to six decimal places, though answers with absolute error of up to  are acceptable.

Example

There are  n=5 elements,  two positive, two negative and one zero. Their ratios are 2/5,2/5  and 1/5 . Results are printed as:

0.400000
0.400000
0.200000


PYTHON FORMATING to 6 decimal places.    print ("{0:.6f}".format(k/n)) 
To know whether number is positive, negative or zero
  for i in range(n):
        if arr[i]>0:
           k=k+1
        elif arr[i]<0:
              l=l+1
        else:
              m=m+1